+++ Client/package.json
@@ -14,14 +14,8 @@
     "@react-navigation/native": "^5.9.4",
     "@react-navigation/stack": "^5.14.5",
     "axios": "^0.21.1",
-<<<<<<< HEAD
-    "bcryptjs": "^2.4.3",
-    "expo": "~40.0.0",
-    "expo-screen-orientation": "~2.1.0",
-=======
     "expo": "^41.0.1",
     "expo-screen-orientation": "~3.1.0",
->>>>>>> 5d35f4279ab3b239401783423ceee9aca0a58716
     "expo-status-bar": "~1.0.3",
     "i": "^0.3.6",
     "react": "16.13.1",
@@ -38,10 +32,6 @@
     "react-native-webview": "11.2.3",
     "react-redux": "^7.2.4",
     "redux": "^4.1.0",
-<<<<<<< HEAD
-    "redux-logger": "^3.0.6",
-=======
->>>>>>> 5d35f4279ab3b239401783423ceee9aca0a58716
     "redux-saga": "^1.1.3"
   },
   "devDependencies": {
