+++ Client/src/navigation/index.js
@@ -3,9 +3,9 @@ import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import { FontAwesome, MaterialIcons } from 'react-native-vector-icons';
-import Signup from '../screens/Signup';
-import Signup2 from '../screens/Signup-Step2';
-import Login from '../screens/Login';
+import Signup from '../screens/Account/Signup/index';
+import Signup2 from '../screens/Account/Signup-Step2';
+import Login from '../screens/Account/Login';
 import Message from '../screens/Home/Message';
 import Community from '../screens/Home/Community';
 import Personal from '../screens/Home/Personal';
